pyproject.toml
[project]
name = "sgs-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "flask>=3.1.0",
    "graphviz>=0.20.3",
    "h5py>=3.13.0",
    "julia>=0.6.2",
    "matplotlib>=3.10.3",
    "mmh3>=5.1.0",
    "networkx>=3.5",
    "notebook>=7.4.0",
    "numpy>=2.2.3",
    "openai>=1.69.0",
    "pandas>=2.2.3",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
    "starlette>=0.46.0",
    "tensorly>=0.9.0",
    "torch>=2.8.0",
    "uvicorn>=0.34.0",
    "mysql-connector-python>=9.0.0",
    "python-multipart>=0.0.9",
    "anthropic>=0.49.0",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "python-docx>=1.1.0",
    "pdfplumber>=0.11.0",
]
